Columns with same names overrides value
On PostgreSQL, when executing one query that returns a column name more
than once, the last column overrides the previous values.
Ex: SELECT id, 99 AS id FROM user
When we got the result, the row has a single property called 'id', and
it's value is 99, overriding the real user ID.
To fix this bug we switched to 'array' rowMode.