Switch react Component to PureComponent
Switched react components to extends PureComponent instead of Component.
This change was necessary due to when we have a tree of Component, and a
parent change its state, all children components are re-rendered, even if
the children attributes arent changed. Switching to PureComponent enables
React to verify if the children components really need to be re-rendered.
Using Component was causing performance issues when the result table is
too large, and we type something on query editor, causing a slow down,
because all result table was being re-rendered.